Update phone/phone.html
removed self closure tag. Login was not working with it.
Reference: http://www.w3.org/TR/xhtml1/#C_3

Given an empty instance of an element whose content model is not EMPTY (for example, an empty title or paragraph) do not use the minimized form (e.g. use <p> </p> and not <p />).